pyproject.toml
[project]
name = "wallet-bot"
version = "0.0.0"
description = "Telegram bot that converts ticket screenshots into Google Wallet passes"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "python-telegram-bot>=21.0,<22",
    "pydantic-settings>=2.0",
    "google-genai>=0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "ruff>=0.5",
    "httpx>=0.27",
    "asgi-lifespan>=2.1",
]
# eval: throwaway benchmarking tools — not installed in prod, not required for tests.
# surya-ocr is intentionally excluded here: it pulls in PyTorch + CUDA (>2 GB) and
# is skipped gracefully by the script when not available.
# To try surya: pip install surya-ocr  (outside Docker, or in a separate image).
eval = [
    "pdfplumber>=0.11",
    "pytesseract>=0.3",
    "Pillow>=10",
    "pdf2image>=1.17",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
# B008: false-positive for FastAPI's Depends() in default arguments.
ignore = ["B008"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-ra"
testpaths = ["tests"]
